+++ shared/bridge/src/commonMain/kotlin/SplashScreen.kt
@@ -1,3 +1,6 @@
+import androidx.compose.runtime.Composable
+
+@Composable
 fun SplashScreen() {
 
 }
