+++ shared/core-utils/src/commonMain/kotlin/extensions/StateFlowExtensions.kt
@@ -1,0 +1,10 @@
+package extensions
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.State
+import kotlinx.coroutines.flow.StateFlow
+import kotlin.coroutines.CoroutineContext
+import kotlin.coroutines.EmptyCoroutineContext
+
+@Composable
+expect fun <T> StateFlow<T>.collectAsStateMultiPlatform(context: CoroutineContext = EmptyCoroutineContext): State<T>
