+++ shared/core/src/androidMain/kotlin/database/DbDriverFactory.kt
@@ -5,10 +5,12 @@ import PlatformConfiguration
 import app.cash.sqldelight.db.SqlDriver
 import app.cash.sqldelight.driver.android.AndroidSqliteDriver
 import card.business.Database
+import org.koin.core.component.KoinComponent
+import org.koin.core.component.inject
 
-actual class DriverFactory(
-    private val platformConfiguration: PlatformConfiguration
-) {
+actual class DriverFactory : KoinComponent {
+
+    private val platformConfiguration: PlatformConfiguration by inject()
 
     actual fun createDriver(): SqlDriver = AndroidSqliteDriver(
         Database.Schema,
