+++ shared/core/src/commonMain/kotlin/database/DbDriverFactory.kt
@@ -3,7 +3,7 @@ package database
 import app.cash.sqldelight.db.SqlDriver
 import card.business.Database
 
-expect class DriverFactory {
+expect class DriverFactory() {
     fun createDriver(): SqlDriver
 }
