+++ shared/core/src/commonMain/kotlin/database/di/DatabaseModule.kt
@@ -1,0 +1,14 @@
+package database.di
+
+import card.business.Database
+import database.DriverFactory
+import org.koin.core.module.dsl.singleOf
+import org.koin.dsl.module
+
+val databaseModule = module {
+    singleOf(DriverFactory::createDriver)
+    singleOf(::DriverFactory)
+    single {
+        Database(driver = get())
+    }
+}
