+++ shared/core/src/commonMain/kotlin/di/CoreModule.kt
@@ -1,4 +1,7 @@
+package di
+
 import coroutines.di.coroutinesModule
+import database.di.databaseModule
 import json.serializationModule
 import ktor.di.ktorModule
 import org.koin.dsl.module
@@ -7,6 +10,7 @@ val coreModule = module {
     includes(
         serializationModule,
         coroutinesModule,
-        ktorModule
+        ktorModule,
+        databaseModule
     )
 }
