+++ shared/core/src/commonMain/kotlin/di/Koin.kt
@@ -1,13 +1,12 @@
 package di
 
-import coreModule
 import org.koin.core.KoinApplication
 import org.koin.core.context.startKoin
 import org.koin.core.module.Module
 
 fun initKoin(vararg exportedModules: Module): KoinApplication = startKoin {
     modules(
+        coreModule,
         *exportedModules,
-        coreModule
     )
 }
