+++ shared/features/employees/api/src/commonMain/kotlin/dto/EmployeeDtoWrapper.kt
@@ -1,0 +1,6 @@
+package dto
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class EmployeeDtoWrapper(val employees: List<EmployeeDto>)
