+++ shared/features/employees/api/src/commonMain/kotlin/network/EmployeesApi.kt
@@ -1,6 +1,7 @@
 package network
 
 import dto.EmployeeDto
+import dto.EmployeeDtoWrapper
 
 interface EmployeesApi {
