+++ shared/features/employees/api/src/commonMain/kotlin/usecases/ListenEmployees.kt
@@ -1,10 +1,12 @@
 package usecases
 
+import kotlinx.coroutines.flow.Flow
+import models.Employee
 import repositories.EmployeeRepository
 
 class ListenEmployees(
     private val employeeRepository: EmployeeRepository
 ) {
 
-    operator fun invoke() = employeeRepository.listenEmployees()
+    operator fun invoke(): Flow<List<Employee>> = employeeRepository.listenEmployees()
 }
